added test setup fixture
