+++ tests/test_deploy_github_pages.py
@@ -3,133 +3,92 @@ from subprocess import run
 from tempfile import TemporaryDirectory
 import os
 import exasol_sphinx_github_pages_generator.deploy_github_pages as deploy_github_pages
+from helper_test_functions import remove_branch, setup_workdir
+from fixtures import setup_test_env
 
 # TODO change to oauth2?
 
 
+def test_remote_branch_creation(setup_test_env):
+    _, _, cwd = setup_test_env
+    source_branch = "5-add-tests"
+    run(["git", "checkout", source_branch], check=True)
+    target_branch = "test-docu-new-branch"
+    remove_branch(target_branch)
 
-def remove_branch(branch_name):
-    target_branch_exists = run(["git", "show-branch", f"remotes/origin/{branch_name}"], capture_output=True,
+    deploy_github_pages.deploy_github_pages(["--target_branch", target_branch,
+                                             "--push_origin", "origin",
+                                             "--push_enabled", "push",
+                                             "--source_branch", source_branch,
+                                             "--source_dir", cwd,
+                                             "--module_path", ["../test_package", "../another_test_package"]])
+    target_branch_exists = run(["git", "show-branch", f"remotes/origin/{target_branch}"], capture_output=True,
                                text=True)
-    # remove remote target branch if exists:
-    if target_branch_exists.returncode == 0:
-        run(["git", "push", "-d", "origin", branch_name], check=True)
+    assert target_branch_exists.returncode == 0
+    remove_branch(target_branch)
 
 
-def setup_test_repo():
-    user_access_token = os.environ.get("MAuserPAT")
-    user_name = os.environ.get("MAuserName")
-    run(["git", "clone", f"https://{user_access_token}@github.com/exasol/sphinx-github-pages-generator-test.git"],
-        check=True)
-    os.chdir("sphinx-github-pages-generator-test")
-    run(["git", "remote", "set-url", "origin",
-         f"https://{user_name}:{user_access_token}@github.com/exasol/sphinx-github-pages-generator-test.git"],
-        check=True)
+def test_pushing_to_existing_docu_branch_same_source(setup_test_env):
+    user_name, user_access_token, cwd = setup_test_env
+    source_branch = "5-add-tests"
+    run(["git", "checkout", source_branch], check=True)
+    target_branch = "test-docu-new-branch"
+    remove_branch(target_branch)
 
-    run(["git", "config", "--local", "user.email", f"{user_name}@exasol.com"], check=True)
-    run(["git", "config", "--local", "user.name", user_name], check=True)
-    return user_name, user_access_token
+    deploy_github_pages.deploy_github_pages(["--target_branch", target_branch,
+                                             "--push_origin", "origin",
+                                             "--push_enabled", "push",
+                                             "--source_branch", source_branch,
+                                             "--source_dir", cwd,
+                                             "--module_path", ["../test_package", "../another_test_package"]])
+    current_commit_id = run(["git", "ls-remote",
+                             f"https://{user_access_token}@github.com/exasol/sphinx-github-pages-generator-test.git",
+                             target_branch], capture_output=True, check=True)
+    commit_id_old = current_commit_id.stdout
 
+    # make a change in the docu
+    with open("./index.rst", "a") as file:
+        file.write("\n\nThis text is a change.")
 
-def test_remote_branch_creation():
-    with TemporaryDirectory() as tempdir:
-        os.chdir(tempdir)
-        setup_test_repo() # todo make into fixture?
-        # change into the "tmpXXX/../branch_name/doc" directory. This is necessary for Sphinx.
-        # the [:-1] removes the newline from the output
-        doc_dir = run(["git", "rev-parse", "--show-toplevel"], capture_output=True, text=True, check=True)
-        os.chdir(f"{doc_dir.stdout[:-1]}/doc")
-        source_branch = "5-add-tests"
-        run(["git", "checkout", source_branch], check=True)
-        target_branch = "test-docu-new-branch"
-        remove_branch(target_branch)
+    deploy_github_pages.deploy_github_pages(["--target_branch", target_branch,
+                                             "--push_origin", "origin",
+                                             "--push_enabled", "push",
+                                             "--source_branch", source_branch,
+                                             "--source_dir", cwd,
+                                             "--module_path", ["../test_package", "../another_test_package"]])
+    current_commit_id = run(["git", "ls-remote",
+                             f"https://{user_access_token}@github.com/exasol/sphinx-github-pages-generator-test.git",
+                             target_branch], capture_output=True, check=True)
+    commit_id_new = current_commit_id.stdout
 
-        cwd = os.getcwd()
-        deploy_github_pages.deploy_github_pages(["--target_branch", target_branch,
-                                                 "--push_origin", "origin",
-                                                 "--push_enabled", "push",
-                                                 "--source_branch", source_branch,
-                                                 "--source_dir", cwd,
-                                                 "--module_path", ["../test_package", "../another_test_package"]])
-        target_branch_exists = run(["git", "show-branch", f"remotes/origin/{target_branch}"], capture_output=True,
-                                   text=True)
-        assert target_branch_exists.returncode == 0
-        remove_branch(target_branch)
-
-
-def test_pushing_to_existing_docu_branch_same_source():
-    with TemporaryDirectory() as tempdir:
-        os.chdir(tempdir)
-        user_name, user_access_token = setup_test_repo()
-        doc_dir = run(["git", "rev-parse", "--show-toplevel"], capture_output=True, text=True, check=True)
-        os.chdir(f"{doc_dir.stdout[:-1]}/doc")
-        source_branch = "5-add-tests"
-        run(["git", "checkout", source_branch], check=True)
-        target_branch = "test-docu-new-branch"
-        remove_branch(target_branch)
-
-        cwd = os.getcwd()
-        deploy_github_pages.deploy_github_pages(["--target_branch", target_branch,
-                                                 "--push_origin", "origin",
-                                                 "--push_enabled", "push",
-                                                 "--source_branch", source_branch,
-                                                 "--source_dir", cwd,
-                                                 "--module_path", ["../test_package", "../another_test_package"]])
-        current_commit_id = run(["git", "ls-remote",
-                                 f"https://{user_access_token}@github.com/exasol/sphinx-github-pages-generator-test.git",
-                                 target_branch], capture_output=True, check=True)
-        commit_id_old = current_commit_id.stdout
-
-        # make a change in the docu
-        with open("./index.rst", "a") as file:
-            file.write("\n\nThis text is a change.")
-
-        deploy_github_pages.deploy_github_pages(["--target_branch", target_branch,
-                                                 "--push_origin", "origin",
-                                                 "--push_enabled", "push",
-                                                 "--source_branch", source_branch,
-                                                 "--source_dir", cwd,
-                                                 "--module_path", ["../test_package", "../another_test_package"]])
-        current_commit_id = run(["git", "ls-remote",
-                                 f"https://{user_access_token}@github.com/exasol/sphinx-github-pages-generator-test.git",
-                                 target_branch], capture_output=True, check=True)
-        commit_id_new = current_commit_id.stdout
-
-        # the docu files where updated, so a new commit should be pushed to the remote
-        assert not commit_id_new == ""
-        assert not commit_id_old == ""
-        assert not commit_id_old == commit_id_new
-        remove_branch(target_branch)
+    # the docu files where updated, so a new commit should be pushed to the remote
+    assert not commit_id_new == ""
+    assert not commit_id_old == ""
+    assert not commit_id_old == commit_id_new
+    remove_branch(target_branch)
 
 
-def test_pushing_to_existing_docu_branch_different_source():
-    with TemporaryDirectory() as tempdir:
-        os.chdir(tempdir)
-        setup_test_repo()
-        doc_dir = run(["git", "rev-parse", "--show-toplevel"], capture_output=True, text=True, check=True)
-        os.chdir(f"{doc_dir.stdout[:-1]}/doc")
-        source_branch_one = "5-add-tests"
-        run(["git", "checkout", source_branch_one], check=True)
-        target_branch = "test-docu-new-branch"
-        remove_branch(target_branch)
+def test_pushing_to_existing_docu_branch_different_source(setup_test_env):
+    _, _, cwd = setup_test_env
 
-        cwd = os.getcwd()
-        deploy_github_pages.deploy_github_pages(["--target_branch", target_branch,
-                                                 "--push_origin", "origin",
-                                                 "--push_enabled", "push",
-                                                 "--source_branch", source_branch_one,
-                                                 "--source_dir", cwd,
-                                                 "--module_path", ["../test_package", "../another_test_package"]])
+    source_branch_one = "5-add-tests"
+    run(["git", "checkout", source_branch_one], check=True)
+    target_branch = "test-docu-new-branch"
+    remove_branch(target_branch)
+
+    deploy_github_pages.deploy_github_pages(["--target_branch", target_branch,
+                                             "--push_origin", "origin",
+                                             "--push_enabled", "push",
+                                             "--source_branch", source_branch_one,
+                                             "--source_dir", cwd,
+                                             "--module_path", ["../test_package", "../another_test_package"]])
 
     with TemporaryDirectory() as tempdir2:
         os.chdir(tempdir2)
-        setup_test_repo()
-        doc_dir = run(["git", "rev-parse", "--show-toplevel"], capture_output=True, text=True, check=True)
-        os.chdir(f"{doc_dir.stdout[:-1]}/doc")
+        _, _, cwd = setup_workdir()
         source_branch_two = "refactoring/1-Move-Sphinx-Documentation-scripts"
         run(["git", "checkout", source_branch_two], check=True)
 
-        cwd = os.getcwd()
         deploy_github_pages.deploy_github_pages(["--target_branch", target_branch,
                                                  "--push_origin", "origin",
                                                  "--push_enabled", "push",
@@ -148,76 +107,64 @@ def test_pushing_to_existing_docu_branch_different_source():
         remove_branch(target_branch)
 
 
-def test_no_new_push_and_commit_if_no_changes():
-    with TemporaryDirectory() as tempdir:
-        os.chdir(tempdir)
-        user_name, user_access_token = setup_test_repo()
-        doc_dir = run(["git", "rev-parse", "--show-toplevel"], capture_output=True, text=True, check=True)
-        os.chdir(f"{doc_dir.stdout[:-1]}/doc")
-        source_branch = "5-add-tests"
-        run(["git", "checkout", source_branch], check=True)
-        target_branch = "test-docu-new-branch"
-        remove_branch(target_branch)
+def test_no_new_push_and_commit_if_no_changes(setup_test_env):
+    user_name, user_access_token, cwd = setup_test_env
+    source_branch = "5-add-tests"
+    run(["git", "checkout", source_branch], check=True)
+    target_branch = "test-docu-new-branch"
+    remove_branch(target_branch)
 
-        cwd = os.getcwd()
-        deploy_github_pages.deploy_github_pages(["--target_branch", target_branch,
-                                                 "--push_origin", "origin",
-                                                 "--push_enabled", "push",
-                                                 "--source_branch", source_branch,
-                                                 "--source_dir", cwd,
-                                                 "--module_path", ["../test_package", "../another_test_package"]])
-        current_commit_id = run(["git", "ls-remote", f"https://{user_access_token}@github.com/exasol/sphinx-github-pages-generator-test.git", target_branch], capture_output=True, check=True)
-        commit_id_old = current_commit_id.stdout
+    deploy_github_pages.deploy_github_pages(["--target_branch", target_branch,
+                                             "--push_origin", "origin",
+                                             "--push_enabled", "push",
+                                             "--source_branch", source_branch,
+                                             "--source_dir", cwd,
+                                             "--module_path", ["../test_package", "../another_test_package"]])
+    current_commit_id = run(["git", "ls-remote", f"https://{user_access_token}@github.com/exasol/sphinx-github-pages-generator-test.git", target_branch], capture_output=True, check=True)
+    commit_id_old = current_commit_id.stdout
 
-        deploy_github_pages.deploy_github_pages(["--target_branch", target_branch,
-                                                 "--push_origin", "origin",
-                                                 "--push_enabled", "push",
-                                                 "--source_branch", source_branch,
-                                                 "--source_dir", cwd,
-                                                 "--module_path", ["../test_package", "../another_test_package"]])
-        current_commit_id = run(["git", "ls-remote", f"https://{user_access_token}@github.com/exasol/sphinx-github-pages-generator-test.git", target_branch], capture_output=True, check=True)
-        commit_id_new = current_commit_id.stdout
+    deploy_github_pages.deploy_github_pages(["--target_branch", target_branch,
+                                             "--push_origin", "origin",
+                                             "--push_enabled", "push",
+                                             "--source_branch", source_branch,
+                                             "--source_dir", cwd,
+                                             "--module_path", ["../test_package", "../another_test_package"]])
+    current_commit_id = run(["git", "ls-remote", f"https://{user_access_token}@github.com/exasol/sphinx-github-pages-generator-test.git", target_branch], capture_output=True, check=True)
+    commit_id_new = current_commit_id.stdout
 
-        # the docu files where not updated, so no new commit should be pushed to the remote
-        assert commit_id_old != 0
-        assert commit_id_old == commit_id_new
-        assert not commit_id_new == ""
-        remove_branch(target_branch)
+    # the docu files where not updated, so no new commit should be pushed to the remote
+    assert commit_id_old != 0
+    assert commit_id_old == commit_id_new
+    assert not commit_id_new == ""
+    remove_branch(target_branch)
 
 
-def test_verify_existence_of_generated_files_on_remote_after_push():
+def test_verify_existence_of_generated_files_on_remote_after_push(setup_test_env):
+    _, _, cwd = setup_test_env
     # generate files locally in test, and with generator, look at diff between local files and remote files?
-    with TemporaryDirectory() as tempdir:
-        os.chdir(tempdir)
-        setup_test_repo()
-        doc_dir = run(["git", "rev-parse", "--show-toplevel"], capture_output=True, text=True, check=True)
-        os.chdir(f"{doc_dir.stdout[:-1]}/doc")
-        source_branch = "5-add-tests"
-        run(["git", "checkout", source_branch], check=True)
-        target_branch = "test-docu-new-branch"
-        remove_branch(target_branch)
-        source_dir = os.getcwd()
-        module_path = ["../test_package", "../another_test_package"]
-        deploy_github_pages.deploy_github_pages(["--target_branch", target_branch,
-                                                 "--push_origin", "origin",
-                                                 "--push_enabled", "push",
-                                                 "--source_branch", source_branch,
-                                                 "--source_dir", source_dir,
-                                                 "--module_path", module_path])
 
-        # __________________
-    with TemporaryDirectory() as tempdir:
-        os.chdir(tempdir)
-        user_name, user_access_token = setup_test_repo()
-        doc_dir = run(["git", "rev-parse", "--show-toplevel"], capture_output=True, text=True, check=True)
-        os.chdir(f"{doc_dir.stdout[:-1]}/doc")
+    source_branch = "5-add-tests"
+    run(["git", "checkout", source_branch], check=True)
+    target_branch = "test-docu-new-branch"
+    remove_branch(target_branch)
+    module_path = ["../test_package", "../another_test_package"]
+    deploy_github_pages.deploy_github_pages(["--target_branch", target_branch,
+                                             "--push_origin", "origin",
+                                             "--push_enabled", "push",
+                                             "--source_branch", source_branch,
+                                             "--source_dir", cwd,
+                                             "--module_path", module_path])
+
+    with TemporaryDirectory() as tempdir2:
+        os.chdir(tempdir2)
+        user_name, user_access_token, cwd = setup_workdir()
         source_branch = "5-add-tests"
         run(["git", "checkout", source_branch], check=True)
-        source_dir = os.getcwd()
-        build_dir = tempdir + "/build"
+        build_dir = tempdir2 + "/build"
+        intermediate_dir = tempdir2 + "/intermediate"
         for module in module_path:
             run(["sphinx-apidoc", "-T", "-e", "-o", "api", module], check=True)
-        run(["sphinx-build", "-b", "html", "-W", source_dir, build_dir], check=True)
+        run(["sphinx-build", "-b", "html", "-d", intermediate_dir, "-W", cwd, build_dir], check=True)
         from pathlib import Path
         from shutil import rmtree
         output_dir = Path("./")
@@ -238,37 +185,32 @@ def test_verify_existence_of_generated_files_on_remote_after_push():
         assert "nothing to commit, working tree clean" in status.stdout
 
 
-def test_only_commit_dont_push():
-    with TemporaryDirectory() as tempdir:
-        os.chdir(tempdir)
-        user_name, user_access_token = setup_test_repo()
-        doc_dir = run(["git", "rev-parse", "--show-toplevel"], capture_output=True, text=True, check=True)
-        os.chdir(f"{doc_dir.stdout[:-1]}/doc")
-        source_branch = "5-add-tests"
-        run(["git", "checkout", source_branch], check=True)
-        target_branch = "test-docu-new-branch"
-        remove_branch(target_branch)
+def test_only_commit_dont_push(setup_test_env):
+    user_name, user_access_token, cwd = setup_test_env
+    source_branch = "5-add-tests"
+    run(["git", "checkout", source_branch], check=True)
+    target_branch = "test-docu-new-branch"
+    remove_branch(target_branch)
 
-        cwd = os.getcwd()
-        deploy_github_pages.deploy_github_pages(["--target_branch", target_branch,
-                                                 "--push_origin", "origin",
-                                                 "--push_enabled", "commit",
-                                                 "--source_branch", source_branch,
-                                                 "--source_dir", cwd,
-                                                 "--module_path", ["../test_package", "../another_test_package"]])
-        current_remote_commit_id = run(["git", "ls-remote", f"https://{user_access_token}@github.com/exasol/sphinx-github-pages-generator-test.git", target_branch], capture_output=True, check=True)
-        remote_commit_id_new = current_remote_commit_id.stdout
+    deploy_github_pages.deploy_github_pages(["--target_branch", target_branch,
+                                             "--push_origin", "origin",
+                                             "--push_enabled", "commit",
+                                             "--source_branch", source_branch,
+                                             "--source_dir", cwd,
+                                             "--module_path", ["../test_package", "../another_test_package"]])
+    current_remote_commit_id = run(["git", "ls-remote", f"https://{user_access_token}@github.com/exasol/sphinx-github-pages-generator-test.git", target_branch], capture_output=True, check=True)
+    remote_commit_id_new = current_remote_commit_id.stdout
 
-        current_local_commit_id = run(["git", "rev-parse", target_branch], capture_output=True, check=True)
-        # the docu files where not updated, so no new commit should be pushed to the remote
-        assert not remote_commit_id_new == current_local_commit_id.stdout
-        assert not current_local_commit_id.stdout == ""
-        target_branch_exists = run(["git", "show-branch", f"remotes/origin/{target_branch}"], capture_output=True,
-                                   text=True)
-        # target branch was not committed, so should not exist on remote
-        assert not target_branch_exists.returncode == 0
+    current_local_commit_id = run(["git", "rev-parse", target_branch], capture_output=True, check=True)
+    # the docu files where not updated, so no new commit should be pushed to the remote
+    assert not remote_commit_id_new == current_local_commit_id.stdout
+    assert not current_local_commit_id.stdout == ""
+    target_branch_exists = run(["git", "show-branch", f"remotes/origin/{target_branch}"], capture_output=True,
+                               text=True)
+    # target branch was not committed, so should not exist on remote
+    assert not target_branch_exists.returncode == 0
 
-        remove_branch(target_branch)
+    remove_branch(target_branch)
 
 
 def test_selection_of_source_branch():
